requirements.txt
aiohttp==3.8.4
discord.py==2.2.2
pandas==1.5.3
psycopg2==2.9.5
pytz==2022.7.1
requests==2.26.0
yt_dlp==2023.3.4
